moved load function to Core/LoadJSON.swift for use in SDK renamed TintButton to AccentButtonStyle and move to file TintedButtonStyle.swift
